refactoring und Zwischenstufe für aufteilung des Codes
