+++ auslesen/auslesen.py
@@ -2,7 +2,7 @@ import sys
 import os
 import logging
 
-import  auslesen.readOutQuestionaires as readOutQuestionaires
+from .readOutQuestionaires import fbs_vorbereiten, phase1_durchfuehren, daten_speichern
 
 logging.basicConfig(filename=f'{(__file__)[:-3]}.log', level=logging.DEBUG, format='%(asctime)s:%(levelname)s:%(message)s', datefmt = "%H:%M-%x")
 logger = logging.getLogger()
@@ -25,9 +25,9 @@ def path_vorbereiten(path):
     return path
 
 def auslesen_starten(path):
-    fbs_liste = readOutQuestionaires.fbs_vorbereiten(path)
-    fbs_fertig_liste = readOutQuestionaires.phase1_durchfuehren(fbs_liste)
-    readOutQuestionaires.daten_speichern(fbs_fertig_liste)
+    fbs_liste = fbs_vorbereiten(path)
+    fbs_fertig_liste = phase1_durchfuehren(fbs_liste)
+    daten_speichern(fbs_fertig_liste)
 
 if __name__ == "__main__":
     path = path_auslesen()
