+++ test_ergs.txt
@@ -1,0 +1,2 @@
+42;0001;1,11;2,1;3,1,5;4,2,8;5,4,12;6,3;7,5;8,4;9,5;10,4;11,5;12,5;13,5;14,3;15,2;16,5;17,5;18,5;19,5;20,4;21,4;22,3;23,2;24,5;25,5;26,5;27,4;28,1;29,3;30,3;31,4;32,3;33,1;34,4;35,3;36,3;37,5;38,5
+42;0002;39,1;40,3;41,5;42,3;43,5;44,1;45,1;46,1,4,7;47,3;48,3;49,3;50,3;51,4;52,3;53,3;54,1;55,1;56,1,4;57,5;58,4;59,3;60,4;61,5;62,3
